add group cluster data display to lighterprints
more refactoring needed to clarify and integrate GroupScores <->
LighterprintsViewModel responsibilities and logic.

+++ fabfile.py
@@ -198,9 +198,11 @@ def deploy(vcs_branch_dict):
                 'chmod -R g+rw logs/',
                 user=env.deploy_user, pty=True)
             env.static_root = vcweb_settings.STATIC_ROOT
-            _virtualenv(sudo, 'pip install -Ur requirements.txt', user=env.deploy_user)
+            if confirm("Update pip dependencies?"):
+                _virtualenv(sudo, 'pip install -Ur requirements.txt', user=env.deploy_user)
+            if confirm("Update cron?"):
+                _virtualenv(sudo, '%(python)s manage.py installtasks' % env, user=env.deploy_user)
             _virtualenv(sudo, '%(python)s manage.py collectstatic' % env, user=env.deploy_user)
-            _virtualenv(sudo, '%(python)s manage.py installtasks' % env, user=env.deploy_user)
             sudo_chain(
                 'chmod -R ug+rw .',
                 'find %(static_root)s %(virtualenv_path)s -type d -exec chmod a+x {} \;' % env,
